+++ source/GameInterface/Services/Time/Patches/DisableGameMenuPausePatches.cs
@@ -56,57 +56,3 @@ internal static class DisableGameMenuPausePatches
     }
 }
 
-[HarmonyPatch(typeof(GameStateManager))]
-class GameStateManagerPatches
-{
-    private static readonly MethodInfo MapState_OnTick = typeof(MapState).GetMethod("OnTick", BindingFlags.NonPublic | BindingFlags.Instance);
-
-    // Prevent pausing in menus without their own game state (such as the encyclopedia)
-    [HarmonyPatch(nameof(GameStateManager.RegisterActiveStateDisableRequest))]
-    static bool Prefix() => false;
-
-
-    // Prevent pausing in menus with their own game states (such as the banner editor, party screen, clan screen, etc.)
-    [HarmonyPatch(nameof(GameStateManager.OnTick))]
-    static void Prefix(ref GameStateManager __instance, float dt)
-    {
-        if (!(__instance.ActiveState is MapState))
-        {
-            MapState mapState = __instance.LastOrDefault<MapState>();
-            if (mapState == null) return;
-
-            MapState_OnTick.Invoke(mapState, new object[] { dt });
-        }
-    }
-
-}
-
-[HarmonyPatch(typeof(MapState), "OnMapModeTick")]
-class MapStatePatch
-{
-    // Remove "base.GameStateManager.ActiveState == this" condition from if statement
-    // to enable campaign ticks while in menu states
-    private static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions)
-    {
-        List<CodeInstruction> instrs = instructions.ToList();
-
-        MethodInfo activeStateGetter = typeof(GameStateManager).GetProperty(nameof(GameStateManager.ActiveState)).GetGetMethod();
-
-        for (int i = 0; i < instrs.Count(); i++)
-        {
-            var instr = instrs[i];
-
-            if (instr.opcode == OpCodes.Callvirt &&
-               instr.operand as MethodInfo == activeStateGetter &&
-               instrs.Count() > i + 2 && 
-               instrs[i - 2].opcode == OpCodes.Ldarg_0 &&
-               instrs[i + 2].opcode == OpCodes.Bne_Un_S)
-            {
-                instrs.RemoveRange(i - 2, 5);
-                break;
-            }
-        }
-
-        return instrs;
-    }
-}
