+++ source/GameInterface/Services/Time/Patches/GameStateManagerPatches.cs
@@ -1,0 +1,35 @@
+﻿using HarmonyLib;
+using System;
+using System.Collections.Generic;
+using System.Reflection;
+using System.Text;
+using TaleWorlds.CampaignSystem.GameState;
+using TaleWorlds.Core;
+
+namespace GameInterface.Services.Time.Patches
+{
+    [HarmonyPatch(typeof(GameStateManager))]
+    class GameStateManagerPatches
+    {
+        private static readonly MethodInfo MapState_OnTick = typeof(MapState).GetMethod("OnTick", BindingFlags.NonPublic | BindingFlags.Instance);
+
+        // Prevents pausing in menus without their own game state (such as the encyclopedia)
+        [HarmonyPatch(nameof(GameStateManager.RegisterActiveStateDisableRequest))]
+        static bool Prefix() => false;
+
+
+        // Prevents pausing in menus with their own game states (such as the banner editor, party screen, clan screen, etc.)
+        [HarmonyPatch(nameof(GameStateManager.OnTick))]
+        static void Prefix(ref GameStateManager __instance, float dt)
+        {
+            if (!(__instance.ActiveState is MapState))
+            {
+                MapState mapState = __instance.LastOrDefault<MapState>();
+                if (mapState == null) return;
+
+                MapState_OnTick.Invoke(mapState, new object[] { dt });
+            }
+        }
+
+    }
+}
